Se agregan constantes y se hace llamado para obtener un producto y su descripción con llamado a crashlytics en caso de que falle

+++ app/src/main/java/com/example/app_challenge_meli/adapters/ProductsAdapter.kt
@@ -35,8 +35,7 @@ class ProductsAdapter (var products: List<Producto> = emptyList(), private val l
         fun bind(product: Producto) {
             binding.title.text = product.title
             product.thumbnail?.let { binding.thumb.loadUrl(it) }
-            binding.txtPrice.text = product.price?.let { convertPrice(it.toDouble()) }
-            //binding.favorite.setImageResource(if (product.favorite) R.drawable.ic_favorite else 0)
+            binding.txtPrice.text = product.price?.let { convertPrice(it) }
         }
     }
